fix: Session: JSON.string: circular value issue
extracted *.ts modules from syn.ts
