+++ ui/src/Content.ts
@@ -1,0 +1,5 @@
+export interface Content {
+  title:string
+  body:string
+  meta?:Record<string, number>
+}
